# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for GestureGuru",
        "title": "Gesture Guru API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/gesture-guru",
    "paths": {
        "/lesson": {
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Detail Lesson",
                "parameters": [
                    {
                        "description": "LessonBrowse",
                        "name": "lesson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LessonBrowse"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/lesson/attempt/{userId}/{lessonId}": {
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Attempt Lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "lessonId",
                        "name": "lessonId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lesson Attempt",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LessonAttempt"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/lesson/save/{userId}/{lessonId}": {
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Save Lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "lessonId",
                        "name": "lessonId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/lessons": {
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Browse Lesson",
                "parameters": [
                    {
                        "description": "LessonBrowse",
                        "name": "lesson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LessonBrowse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Lesson"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Lesson": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "saved": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
        "models.LessonAttempt": {
            "type": "object",
            "required": [
                "score"
            ],
            "properties": {
                "attemptNumber": {
                    "type": "integer"
                },
                "score": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 1
                }
            }
        },
        "models.LessonBrowse": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "lessonId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}